年月先tなくモーダル、コンポーネント化対応
