+++ src/app/(tabs)/index.tsx
@@ -6,9 +6,8 @@ import { collection, onSnapshot, query, orderBy, where } from 'firebase/firestor
 import { DiaryType } from '../../../type/diary';
 import PlusIcon from '../components/Icon/PlusIcon';
 import { useRouter } from 'expo-router';
-import Modal from "react-native-modal";
-import { Picker } from '@react-native-picker/picker';
 import dayjs from 'dayjs';
+import YearMonthSelectModal from '../diaryList/components/YearMonthSelectModal';
 
 export default function home() {
   const [diaryLists, setDiaryLists] = useState<DiaryType[]>([]);
@@ -23,24 +22,6 @@ export default function home() {
   // 選択された年月を'YYYY-M'形式の文字列で保持するstate
   const [selectedYearMonth, setSelectedYearMonth] = useState(displayDate.format('YYYY-M'));
 
-  // ピッカー用の年月リストを生成する
-  const generateYearMonths = () => {
-    const list = [];
-
-    const startYear = dayjs().year() - 10;
-    const endYear = dayjs().year() + 10;
-    for (let year = endYear; year >= startYear; year--) {
-      // 新しい月から表示するため、12月から1月へループ
-      for (let month = 12; month >= 1; month--) {
-        list.push({
-          label: `${year}年${month}月`,
-          value: `${year}-${month}`
-        });
-      }
-    }
-    return list;
-  };
-  const yearMonths = generateYearMonths();
 
   useEffect(() => {
     const userId = auth.currentUser?.uid;
@@ -68,25 +49,6 @@ export default function home() {
     setModalVisible(true);
   }
 
-  const handleModalClose = () => {
-    setModalVisible(false);
-  }
-
-  const handleModalConfirm = () => {
-    // OKボタンが押されたら表示年月を更新
-    // 'YYYY-M'形式の文字列からdayjsオブジェクトを生成
-    setDisplayDate(dayjs(selectedYearMonth));
-    setModalVisible(false);
-  }
-
-// 「今月」ボタンが押されたときの処理
-  const handleSelectThisMonth = () => {
-    // 現在の年月を'YYYY-M'形式で取得
-    const currentMonth = dayjs().format('YYYY-M');
-    // ピッカーの選択値を現在の年月に更新する
-    setSelectedYearMonth(currentMonth);
-  };
-
   return (
     <View style={styles.container}>
       {/* 年月 */}
@@ -110,34 +72,13 @@ export default function home() {
         <PlusIcon width={30} height={30} color="white" />
       </TouchableOpacity>
       {/* 年月選択モーダル */}
-      <Modal
-        isVisible={isModalVisible}
-        onBackdropPress={handleModalClose} // 背景をタップしたら閉じる
-        style={styles.modal}
-      >
-        <View style={styles.modalContent}>
-          <View style={styles.modalHeader}>
-            <TouchableOpacity onPress={handleModalClose} style={styles.modalHeaderButton}>
-              <Text style={styles.modalButtonText}>キャンセル</Text>
-            </TouchableOpacity>
-            <TouchableOpacity onPress={handleSelectThisMonth} style={styles.modalHeaderButton}>
-              <Text style={[styles.modalButtonText, styles.modalConfirmText]}>今月</Text>
-            </TouchableOpacity>
-            <TouchableOpacity onPress={handleModalConfirm} style={styles.modalHeaderButton}>
-              <Text style={[styles.modalButtonText, styles.modalConfirmText]}>OK</Text>
-            </TouchableOpacity>
-          </View>
-          {/* ピッカー */}
-          <Picker
-            selectedValue={selectedYearMonth}
-            onValueChange={(itemValue) => setSelectedYearMonth(itemValue)}
-          >
-            {yearMonths.map(item => (
-              <Picker.Item key={item.value} label={item.label} value={item.value} />
-            ))}
-          </Picker>
-        </View>
-      </Modal>
+      <YearMonthSelectModal
+        setModalVisible={setModalVisible}
+        setDisplayDate={setDisplayDate}
+        setSelectedYearMonth={setSelectedYearMonth}
+        isModalVisible={isModalVisible}
+        selectedYearMonth={selectedYearMonth}
+      />
     </View>
   )
 }
@@ -147,7 +88,6 @@ const styles = StyleSheet.create({
     flex: 1,
   },
   yearMonthContainer: {
-    // View用のスタイル（必要に応じて）
     backgroundColor: '#ffffff',
   },
   yearMonthText: {
@@ -175,38 +115,6 @@ const styles = StyleSheet.create({
     shadowOpacity: 0.25,
     shadowRadius: 8,
     elevation: 5,
-  },
-  modal: {
-    justifyContent: 'flex-end',
-    margin: 0,
-  },
-  modalContent: {
-    backgroundColor: 'white',
-    borderTopLeftRadius: 20,
-    borderTopRightRadius: 20,
-    paddingHorizontal: 20,
-    paddingBottom: 20, // SafeAreaViewを考慮する場合は調整
-  },
-  modalHeader: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    paddingVertical: 15,
-    borderBottomWidth: 1,
-    borderBottomColor: '#E0E0E0',
-  },
-  modalHeaderButton: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  modalButtonText: {
-    fontSize: 16,
-    color: '#007AFF',
-    fontWeight: 'bold',
-  },
-  modalConfirmText: {
-    fontWeight: 'bold',
   },
   noDiaryText: {
     textAlign: 'center',
